requirements.txt
requests>=2.32.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9

# Web UI dependencies
streamlit>=1.28.0
pandas>=2.0.0
